backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
pydantic>=2.7
pandas>=2.2
numpy>=1.26
requests>=2.32
aiohttp>=3.10
tenacity>=8.3

faiss-cpu>=1.8
sentence-transformers>=3.0

langchain==0.3.27
langchain-core>=0.3.72,<1.0.0
langchain-community==0.3.21
langchain-huggingface<1.0.0
python-dotenv>=1.0    # (선택) .env 파일에서 환경변수 읽고 싶을 때
